Refactor: CamerRepresentable CamerView 파일 통합 CamerRepresentable 파일 삭제
